Modification : changement emplacement fonction calcul duree dans la classe trajet et opérationnel

+++ src/Headers/Domaine.h
@@ -10,6 +10,7 @@
 
 
 
+
 class Domaine {
 private:
     std::map<int, Sommet*> m_sommets;
@@ -27,7 +28,7 @@ public:
     //Méthodes
     void creationSommets(const std::vector<t_chargeSommets>& _som);
     void initialisation(const t_chargeFichier &fCharge);
-    int calculDuree(const std::string nom,const std::string type,int depart, int arrivee);
+    float calculDuree(const std::string nom,const std::string type,int depart, int arrivee);
 
     //Getters & Setters
     void setOrdre(int _ordre);
